requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.3
atari-py==0.2.6
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
future==0.18.2
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.25.0
gym==0.15.4
h5py==2.10.0
idna==2.8
isort==4.3.21
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.17.4
oauthlib==3.1.0
objgraph==3.4.1
opencv-python==4.1.1.26
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pyglet==1.3.2
pylint==2.4.4
pyparsing==2.4.5
python-dateutil==2.8.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.3.2
six==1.13.0
tb-nightly==2.1.0a20191123
tensorboard==2.0.1
tensorflow-estimator==2.0.1
termcolor==1.1.0
tf-estimator-nightly==2.0.0.dev2019112309
tf-nightly-gpu==2.1.0.dev20191123
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
